python/pytorch_server/requirements.txt
# required by '.env' and '.flaskenv' files
python-dotenv~=0.21.1
flask==2.0.2
flask[async]
elasticsearch[async]
flask-cors
flask-wtf==1.0.1
Pillow~=9.3.0
# sentence-transformers~=2.2.2
sentence-transformers
tqdm~=4.64.1
WerkZeug==2.2.2
git+https://github.com/openai/CLIP.git
regex 
ftfy 

# required by '.env' and '.flaskenv' files
python-dotenv~=0.21.1

# for reading the exif jpg data
exif==1.5.0

# required only for RAW images conversion
# rawpy==0.17.3
# imageio==2.22.4

-f https://download.pytorch.org/whl/cu116
torch==1.13.1

